+++ src/Humanizer/Configuration/OrdinalizerRegistry.cs
@@ -6,11 +6,12 @@ namespace Humanizer.Configuration
     {
         public OrdinalizerRegistry() : base(new DefaultOrdinalizer())
         {
+            Register("de", new GermanOrdinalizer());
             Register("en", new EnglishOrdinalizer());
             Register("es", new SpanishOrdinalizer());
-            Register("ru", new RussianOrdinalizer());
+            Register("it", new ItalianOrdinalizer());
             Register("pt-BR", new BrazilianPortugueseOrdinalizer());
-            Register("de", new GermanOrdinalizer());
+            Register("ru", new RussianOrdinalizer());
             Register("tr", new TurkishOrdinalizer());
         }
     }