+++ release_notes.md
@@ -1,4 +1,6 @@
 ###In Development
+  - [#320](https://github.com/MehdiK/Humanizer/pull/320): Fixed Dehumanize actually humanizing an alreaddy dehumanized string
+
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.28.0...master)
 
 ###v1.28.0 - 2014-07-06