+++ release_notes.md
@@ -12,6 +12,7 @@
   - [#307](https://github.com/MehdiK/Humanizer/pull/307): Added support to string.FormatWith for the explicit culture parameter
   - [#312](https://github.com/MehdiK/Humanizer/pull/312): Added Turkish ToWord, ToOrdinalWord and Ordinalize implementation
   - [#173](https://github.com/MehdiK/Humanizer/pull/173): Added support for Window Phone 8.1
+  - [#314](https://github.com/MehdiK/Humanizer/pull/314): Added ByteRate class and supporting members to facilitate calculation of byte transfer rates
 
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.27.0...v1.28.0)